Rename solvers code to follow cppguide
(Does not include fixes in subdirectories, just solvers top-level.)
