+++ drake/examples/Pendulum/pendulum_swing_up.cc
@@ -8,7 +8,7 @@
 #include "drake/examples/Pendulum/gen/pendulum_state_vector.h"
 #include "drake/examples/Pendulum/pendulum_swing_up.h"
 #include "drake/examples/Pendulum/pendulum_plant.h"
-#include "drake/solvers/Function.h"
+#include "drake/solvers/function.h"
 #include "drake/systems/plants/constraint/direct_collocation_constraint.h"
 
 using drake::solvers::detail::VecIn;
