+++ drake/solvers/function.h
