+++ drake/solvers/system_identification.cc
